+++ packages/dynamic-bonding-curve/src/services/creator.ts
@@ -7,6 +7,7 @@ import {
     type Connection,
 } from '@solana/web3.js'
 import {
+    ClaimCreatorTradingFee2Param,
     ClaimCreatorTradingFeeParam,
     ClaimCreatorTradingFeeWithQuoteMintNotSolParam,
     ClaimCreatorTradingFeeWithQuoteMintSolParam,
@@ -124,7 +125,8 @@ export class CreatorService extends DynamicBondingCurveProgram {
                 payer,
                 tokenBaseAccount,
                 feeReceiver,
-                poolState.baseMint
+                poolState.baseMint,
+                tokenBaseProgram
             )
         createTokenBaseAccountIx &&
             preInstructions.push(createTokenBaseAccountIx)
@@ -134,7 +136,8 @@ export class CreatorService extends DynamicBondingCurveProgram {
                 payer,
                 tokenQuoteAccount,
                 tempWSolAcc,
-                poolConfigState.quoteMint
+                poolConfigState.quoteMint,
+                tokenQuoteProgram
             )
         createTokenQuoteAccountIx &&
             preInstructions.push(createTokenQuoteAccountIx)
@@ -309,6 +312,123 @@ export class CreatorService extends DynamicBondingCurveProgram {
     }
 
     /**
+     * Claim creator trading fee
+     * @param claimCreatorTradingFeeParam - The parameters for the claim creator trading fee
+     * @returns A claim creator trading fee transaction
+     */
+    async claimCreatorTradingFee2(
+        claimCreatorTradingFee2Param: ClaimCreatorTradingFee2Param
+    ): Promise<Transaction> {
+        const {
+            creator,
+            pool,
+            maxBaseAmount,
+            maxQuoteAmount,
+            receiver,
+            payer,
+        } = claimCreatorTradingFee2Param
+
+        const poolState = await this.state.getPool(pool)
+
+        if (!poolState) {
+            throw new Error(`Pool not found: ${pool.toString()}`)
+        }
+
+        const poolConfigState = await this.state.getPoolConfig(poolState.config)
+
+        if (!poolConfigState) {
+            throw new Error(`Pool config not found: ${pool.toString()}`)
+        }
+
+        const tokenBaseProgram = getTokenProgram(poolConfigState.tokenType)
+        const tokenQuoteProgram = getTokenProgram(
+            poolConfigState.quoteTokenFlag
+        )
+
+        const isSOLQuoteMint = isNativeSol(poolConfigState.quoteMint)
+
+        if (isSOLQuoteMint) {
+            const preInstructions: TransactionInstruction[] = []
+            const postInstructions: TransactionInstruction[] = []
+
+            const tokenBaseAccount = findAssociatedTokenAddress(
+                receiver,
+                poolState.baseMint,
+                tokenBaseProgram
+            )
+
+            const tokenQuoteAccount = findAssociatedTokenAddress(
+                creator,
+                poolConfigState.quoteMint,
+                tokenQuoteProgram
+            )
+
+            const createTokenBaseAccountIx =
+                createAssociatedTokenAccountIdempotentInstruction(
+                    payer,
+                    tokenBaseAccount,
+                    receiver,
+                    poolState.baseMint,
+                    tokenBaseProgram
+                )
+            createTokenBaseAccountIx &&
+                preInstructions.push(createTokenBaseAccountIx)
+
+            const createTokenQuoteAccountIx =
+                createAssociatedTokenAccountIdempotentInstruction(
+                    payer,
+                    tokenQuoteAccount,
+                    creator,
+                    poolConfigState.quoteMint,
+                    tokenQuoteProgram
+                )
+            createTokenQuoteAccountIx &&
+                preInstructions.push(createTokenQuoteAccountIx)
+
+            const unwrapSolIx = unwrapSOLInstruction(creator, receiver)
+            unwrapSolIx && postInstructions.push(unwrapSolIx)
+
+            const accounts = {
+                poolAuthority: this.poolAuthority,
+                pool,
+                tokenAAccount: tokenBaseAccount,
+                tokenBAccount: tokenQuoteAccount,
+                baseVault: poolState.baseVault,
+                quoteVault: poolState.quoteVault,
+                baseMint: poolState.baseMint,
+                quoteMint: poolConfigState.quoteMint,
+                creator,
+                tokenBaseProgram,
+                tokenQuoteProgram,
+            }
+
+            return this.program.methods
+                .claimCreatorTradingFee(maxBaseAmount, maxQuoteAmount)
+                .accountsPartial(accounts)
+                .preInstructions(preInstructions)
+                .postInstructions(postInstructions)
+                .transaction()
+        } else {
+            const result = await this.claimWithQuoteMintNotSol({
+                creator,
+                payer,
+                feeReceiver: receiver,
+                pool,
+                poolState,
+                poolConfigState,
+                tokenBaseProgram,
+                tokenQuoteProgram,
+            })
+            return this.program.methods
+                .claimCreatorTradingFee(maxBaseAmount, maxQuoteAmount)
+                .accountsPartial(result.accounts)
+                .preInstructions(result.preInstructions)
+                .postInstructions([])
+                .transaction()
+        }
+    }
+
+    /**
      * Withdraw creator surplus
      * @param creatorWithdrawSurplusParam - The parameters for the creator withdraw surplus
      * @returns A creator withdraw surplus transaction
@@ -385,7 +505,8 @@ export class CreatorService extends DynamicBondingCurveProgram {
     ): Promise<Transaction> {
         const { virtualPool, creator, newCreator } = transferPoolCreatorParams
         const virtualPoolState = await this.state.getPool(virtualPool)
-        const migrationMetadata = deriveDammV1MigrationMetadataAddress(virtualPool)
+        const migrationMetadata =
+            deriveDammV1MigrationMetadataAddress(virtualPool)
         const transaction = await this.program.methods
             .transferPoolCreator()
             .accountsPartial({
@@ -406,7 +527,7 @@ export class CreatorService extends DynamicBondingCurveProgram {
         return transaction
     }
 
-    async withdrawMigrationFee(
+    async creatorWithdrawMigrationFee(
         withdrawMigrationFeeParams: WithdrawMigrationFeeParam
     ): Promise<Transaction> {
         const { virtualPool, sender, feePayer } = withdrawMigrationFeeParams